Relevant text from FSF's "GPLv3 Final Rationale" as released on 2007-06-29.
I (Bradley M. Kuhn) went through FSF's "GPLv3 Final Rationale", and pasted in
any sections that seemed useful to this tutorial.  There is a lot of
interesting material in that particular rationale document, although much of
it is probably too verbose for inclusion.  I expect much of this will need to
be cut out.

The raw material used for this commit can be found here:
     http://www.gnu.org/licences/gpl3-final-rationale.pdf

As I pasted in this text, I added FIXME's sometimes where it seemed the text
might need work.  However, I was much more extensive in just pasting here, so
there's a big editing job now.  As mentioned in a previous commit, the whole
GPLv3 chapter is now completely disjoint with all this pasting.

Finally, note that this material was originally copyrighted and licensed as
follows:

  Copyright © 2007, Free Software Foundation, Inc.

  Verbatim copying and distribution of this entire article are permitted
  worldwide, without royalty, in any medium, provided this notice is
  preserved.

However, I am hereby relicensing this material to CC-By-SA-4.0, with the
verbal permission from John Sullivan, Executive Director of the FSF, which
was given to me during a conference call on Wednesday 12 February 2014.  I
also confirmed that relicensing permission on IRC with johnsu01 today.