Remove unnecessary versions and formats of the gpl-lgpl tutorial material.
After the cherry-picking activity just completed in preceeding commits, the
file in the top-level directory named 'gpl-lgpl.tex' now includes a full
history of all the improvements, with proper author dates, names, and emails.

This includes a merger of the old history of 'GPL-Business/gpl-business.tex',
to the middle-period history in the 'GPL-LGPL/gpl-lgpl.tex', and the final
part of history found in  'GPL-LGPL/new-gpl-lgpl.tex'.

Also removed herein are dvi, pdf, and ps files for the new-gpl-lgpl.tex file
that were apparently inadvertently added to the historical CVS repository.