Remove deprecated LaTeX formatting scripts and commands.
These were done by me in 2001 to handle generating HTML, Postscript, and PDF
all from the same LaTeX sources.  Doing this is not as difficult as it once
was, as such, these files are no longer needed to achieve that goal.

However, it's still not automatic, so some additional stuff for formatting to
all types will need to be added later.