Remove unnecessary versions and formats of the case-studiess tutorial material.
After the cherry-picking activity just completed in preceeding commits, the
file in the top-level directory named 'enforcement-case-studies.tex' now
includes a full history of all the improvements, with proper author dates,
names, and emails.

This includes a merger of the old history of
'Case-Studies-Ethics/case-study-ethics.tex', and newer history in
'Case-Studies-Ethics/new-case-study-ethics.tex',

Also removed herein are pdf files for the (new-)case-studies-ethics.tex file
that were apparently inadvertently added to the historical CVS repository.