Properly label license names and proper formatting of section symbol.
Throughout the document, there were inconsistencies in how the \S was used,
and also there were many places where it was used to refer to section numbers
but not the document they were in.

Since the tutorial now covers so many different licenses, be pedantic about
this throughout and ensure the \S is never without its document name.