Relevant text from GPLv3 First Discussion Draft Rationale of 2006-01-16.
I carefully went through FSF's First Discussion Draft Rationale, which was
published on Monday 16 January 2006 and merged in any relevant text and
descriptions that might be of use in this tutorial.

The raw material used for this commit can be found here:
       http://gplv3.fsf.org/gpl-rationale-2006-01-16.html
Specifically, a copy of the LaTeX sources are here:
       http://gplv3.fsf.org/gpl-rationale-2006-01-16.tex

As I merged in this text, I added FIXME's where it seemed the text was
incomplete or referred to parts of GPLv3 draft text that disappeared in later
versions.

Finally, note that this material was originally copyrighted and licensed as
follows:

  Copyright © 2006 Free Software Foundation, Inc.

  Verbatim copying and distribution of this entire article are permitted
  worldwide, without royalty, in any medium, provided this notice, and the
  copyright notice, are preserved.

However, I am hereby relicensing this material to CC-By-SA-4.0, with the
verbal permission from John Sullivan, Executive Director of the FSF, which
was given to me during a conference call on Wednesday 12 February 2014.