Rewrite the entire Patent section.
Much of this text, pasted from various different rationale documents, deals
with issues that either changed in later drafts, or is focused on various
minutia about the Microsoft/Novell deal that seems in retrospect rather
unimportant for many reasons.

Hopefully what I've cut down to is the bare essentials of what's needed to
explain the patent section of GPLv3 to a newbie.  I suspect that I'll want
later to go over this and add more material or reword things further.

I've added a few FIXME-LATER's for that reason.