GPLv3 clarifies private modification permission.
While both GPLv2 and GPLv3 have long been considered to grant unabridged
right to private modification, GPLv3 has much clearer and explicit
wording to this effect.

This should be noted when that paragraph is explained.  This change
herein does that.

HT Wolvereness for pointing this difference between GPLv2 and GPLv3 out
to me.