Explain GPLv2 implicit private modification rights
I've long been aware that GPLv2 "technically" governed private
modifications and that generally there were probably more requirements
on privately modified versions of GPLv2'd works than most people assumed
in practice, including commonly held public interpretation by FSF.

HT Wolvereness, who pointed out to me that GPLv3 solved that problem.
When I spoke to Fontana about it, he was indeed aware that this text was
"missing" in GPLv2 and that GPLv3 properly added it, through some
politics during the GPLv3 process.

I've added herein the ultimate historical conclusions about GPLv2's
interpretation and how GPLv3 clarified it.  I've left out the color
about the politics of how it got added, not because they are not
interesting, relevant and germane to tutorial, but because we don't have
a good place yet in the tutorial for discussion of GPLv3 drafting
politics, and frankly if we have such a section, Fontana ought to write
it, not me.