Move details of exclusive power to a footnote.
enyst was quite correct that more explanation was needed here about how
an entity achieves exclusive relicensing rights.  However, the details
are somewhat off-point to what the section is trying to explain, so the
details are best placed in a footnote.

I've also separated out copyright assignment from generating all of
one's own copyrights.  This may be a distinction without a difference,
but a laundry list seemed appropriate here.  Perhaps this should be
shortened in future.