derivative works matter less than you think.
Given the wealth of text this chapter offers regarding derivative works,
this section really is necessary to give context on this point and
assure that the reader is not unduly swayed to believe that the
derivative work discussion is a central tenant of understanding
copyleft.

Long term, it may make sense to move the entire chapter on derivative
works to a different part of the tutorial.  Historically, it was placed
here because when teaching courses on the subject based on this text, I
found as an instructor that questions about derivative works became so
rampant from students during discussions of GPLv2§2 / GPLv3§5 that
derivative works discussion ahead of time was the only way to quell the
onslaught of ultimately off-topic questions.

Thus, the placement of the derivative works section in this location may
in fact be merely an historical artifact that this text was written
originally to accompany an in-course presentation.  While I'd still
recommend organizing a classroom presentation of these topics in that
order, I no longer believe the written materials must follow suit.