Rework over-abundant use of term "derivative work"
The older portions of this tutorial tended to favor the term "derivative
work", since that was the popular catch-all term used at the time the
text was written.

However, as the newer text regarding GPLv3 now states, FSF abandoned the
use of the term "derivative work" in the text of GPLv3 itself, for
various reasons we already discuss in the tutorial.

Therefore, the tutorial text itself should likely not rely so heavily on
the phrase "derivative work" throughout.  This change herein reworks a
number of places where "derivative work" was used in the tutorial and
replaced it with other terms.

Ultimately, some word-smithing happened as part of the process of doing
this patch.