Add rootfs and U-Boot install notes to case study
This change also fixes the location of the multiple image note, which
is better included after the note added in 3c15418 so that it's clear
what "This step" refers to.

Also added were notes on how we checked to confirm the kernel was
corresponding and commentary on why the toolchain issue was much less
severe than the toolchain issues we usually see.