Move ThinkPenguin CCS case study to first one.
The original text for this case study was developed as the final
chapter, as sort of a "you heard about the rest, now try the best"
approach.  However, I think we want to put this one first.  "Hope before
despair" is perhaps a better attitude?

The diff on this one didn't come out so clean, so it's a bit difficult
to see that I'm just moving  a big block of text up.

I diffed it against a version of this file, from:
  commit 2f2e5f9e4c32e2ea0de694f60b65ec98b4d3958a on 2014-10-15 22:14:45 -0400

that was before work began on this chapter, and the diff was relatively
clean.