Rewrote U-Boot Installation section.
My primary goal here was to put the text into a consistent voice, and
convert the text to a more flowing narrative rather than a step-by-step
list.

In a few places, I added commentary on the process where it seemed
appropriate, but I strove to keep that minimal.

Finally, labels to some unlabeled sections of gpl-lgpl.tex were needed
for back-references used in my rewrite.