Termination inevitability w/ aggregated GPLv2-only
Most redistributors, at least with regard to embedded systems, typically
include the kernel named Linux with their distribution.  As such, even
when GPLv2-or-later and/or GPLv3-{only, or-later} works are included in
the aggregation, the termination implications are effectively those of
GPLv2-only, since it's unlikely in any event the violator will remove
the GPLv2-only work (particularly if it's Linux).