Rework two compliance guide sections in Background
Some of the text pasted in earlier commits was certainly useful, but
needed a complete rework.

Also, the text pasted was far too terse, and more detail was needed.

Therefore, I've moved text around and build a more comprehensive
Background section.  I've moved the burgeoning "Understanding Who's
Enforcing" section into the Background chapter and made it complete.

Probably the most bizarre (?) change I've made here is coining this
acronym COGEO.  This is non-optimal for sure, and I've added a FIXME to
seek a better term.