Link LGPL compliance section to LGPL analysis.
Since so little material is currently given on LGPL compliance, it's
likely best to link back to the chapter on LGPL compliance.

Besides, I don't think there really is anything additional the
compliance guide can add regarding LGPL compliance, other than the
detail license analysis on LGPL already available in that part of the
text.

(Note labels had to be added for the chapters that didn't previously
exist.)