Full rewrite of the new M&A compliance section.
Again, upon careful reading of the pasted text, it was clearly not
as useful as it first appeared, and is in fact somewhat misleading.

This rewrite does a better job explaining the necessary focus required
for an M&A situation.  The section could use some work, but generally
speaking IMO this new does a better job than both the pasted text and
other texts on the issues I've read elsewhere.