Wrote section on "grey hat" GPL enforcement.
The inspiration for this section came from the pasted text, which
ultimately whitewashed this well-known and complex situation.  While my
new text likely has the biases inherent in a COGEO-oriented focused
document, so perhaps future patches that soften that side of it would be
helpful.

However, I believe generally that the new section describes the
situation substantially better than the terse pasted text that lauded
it.

Finally, this section is written to build up to some level of crescendo,
since the conclusion immediately follows it.