Incorporate pasted text re: AGPLv3/GPLv3 combos.
This text was mostly useful as is.  However, it failed to make a key
point I've often made: that the combinations created by comingling
AGPLv3'd code with GPLv3'd code may be difficult to disentangle, and
thus in practice, it may turn out that such a combination effectively
must be licensed as a whole under AGPLv3, even if technically some
copyrights included therein are GPLv3'd.

In practice, this nuance is only a technical barrier, since complying
fully with AGPLv3 automatically complies with GPLv3.