Integrate pasted text regarding LGPL.
Much of the pasted text here was useful.  However, some of the claims
were broad reaching, I've reigned those in.  (e.g., saying "Taken
together, these provisions mean:" was a bit strong).

Also, in that specific spot, the conclusions made in the text were
described as applying to LGPLv2.1, but are clearly conclusions about
LGPLv3.  I've corrected that herein.

Finally, I had to write a bunch of next text to make the pasted text
work, and also added one FIXME for later of where things could be
improved further.