Replace "enforcements" with "enforcement actions".
I think this sounds better.  While "enforcement" is a noun and therefore
it seems it should be "pluralizable", "enforcements" wasn't in my
dictionary as a correctly spelled word, and that made me realize I'd
never heard "enforcements" used in a plural before, and it immediately
sounded weird.  So, I changed it.