Some editorial changes.
Commenting on one: the initial-caps stylistic preference for "Free
Software" (though it contradicts prevailing usage, including that of RMS
and the FSF) ought to be respected, but I think it is confusing to
capitalize the 'S' when referring to nonfree software as "non-Free
Software". So I changed this to "non-Free software" and also implicitly
acknowledged that the preference for "non-Free" over "nonfree" is the
editor-in-chief's stylistic idiosyncrasy.