Replace editorial commentary w/ FSF's link.
My personal comment here, which I wrote on 2003-05-26 (see
f05ce6c657e07a5e6c6def3f7ff8cb2b2bcf6246 ), is probably not particularly
useful.  I still tend to use the phrasing as original stated in the
removed text herein; however, I'm admittedly the only one.  I don't deny
that I hope to coin some terminology usage through my work on this
Guide, but this particular use of "nonfree software" to mean
"noncommercially proprietary" is not so important IMO that this Guide
must coin it.

The FSF's page on this doesn't make that distinction, and has much more
detail on this issue than this section does.  Therefore, the
personal statement is removed, and the organizational statement on the
FSF's site is instead linked to.