Wordsmith sentence after phrase "such software".
I think Fontana's change a few commits back which s/it/such software/ in
this sentence is a useful one.  However, the entire sentence reads even
worse because "such software Free Software" looks so strange as a string
of words in the middle of the sentence.

This change is the best rework I could find that resolved the problem.
It's probably not the best option but is certainly an improvement.