Consolidate copyright notices into single place.
Historically, this project used (more-or-less) a file-by-file copyright
inventory.  This commit ends that practice.  The project now has a
single toplevel copyright inventory, stored exclusively in
comprehensive-gpl-guide.tex (so that it appears also in compiled
versions of the Guide as well).

The side-effect of this commit is that the parts may no longer be easily
publishable separably without (at least) the additional work of
copyright notice reconstruction.  This may in particular create a
challenge for the FSF, who has historically selectively published
sections of this Guide as materials for its CLE classes.

However, without this change, this Guide will eventually suffer from the
inherent problems in maintaining file-by-file copyright inventory.
Circumstances simply dictate a single, top-level copyright and license
notice for the entire Guide.

In addition to consolidation of copyright notices, I've also herein
updated my historical copyright notices to properly credit me for my own
work done in 2003 through 2005.

I've also updated the license notice to reflect the changes made by the
previous commit and related issues.