Clarify who "most people" meant, drop UCITA ref.
The intent of this text was to point out that most users don't actually
believe they get warranties, which is still surely correct, given that
GPL disclaims warranties in the same manner nearly every software
license -- proprietary or free -- does anyway.

Also, the forward-reference to the later section's discussion of UCC
should be hinted at here.  There is no explicit reference to UCC made
here, but it is encompassed in "many local laws", since the later
section mentions the specific section of UCC involved.

Meanwhile, the reference to UCITA is dropped, but perhaps it should be
reintroduced in other text in the main warranty section.  UCITA has had
much less policy impact than was expected when the original version of
this text was written.  It might be useful to ask policy folks and
attorneys from Maryland and Virginia who might be able to help explain
what impact UCITA has had being on the books only there.