Clarify & expound on issue of warranty disclaimer
The existing paragraph on this issue was inadequate, since it punted
entirely to GPLv2§11 for dealing with critics' claims of
unenforceability.  That left a mistaken impression of validity of such
claims.

The commit herein adds reference to CIGS, which likely permits GPL's
sort of warranty disclaimer in most jurisdictions, and also bolsters the
reference to the UCC earlier in the section.

However, given academic debate about the applicability of CIGS to
software licenses, this commit includes a footnote referencing the two
sides of that debate.

Tony Sebro and I co-drafted these changes together.

Signed-Off-By: Tony Sebro <tony@sfconservancy.org>
Signed-Off-By: Bradley M. Kuhn <bkuhn@ebb.org>